+++ index.js
@@ -1,8 +1,9 @@
 'use strict';
 
-const execa = require('execa');
+const childProcess = require('child_process');
 const get = require('lodash.get');
 const { lookpath } = require('lookpath');
+const readline = require('readline');
 
 const GRANULE_LOG_LIMIT = 500;
 
@@ -26,92 +27,55 @@ class CumulusMessageAdapterExecutionError extends Error {
 async function generateCMASpawnArguments(command) {
   const adapterDir = process.env.CUMULUS_MESSAGE_ADAPTER_DIR || './cumulus-message-adapter';
   const systemPython = await lookpath('python');
-  if (systemPython) {
+  if (systemPython && process.env.USE_CMA_BINARY !== 'true') {
     return [systemPython, [`${adapterDir}`, command]];
   }
   // If there is no system python, attempt use of pre-packaged CMA binary
-  return [`${adapterDir}/cma`, [command]];
+  return [`${adapterDir}/cma_bin/cma`, [command]];
 }
+
 
 /**
  * Invoke the cumulus-message-adapter
  *
- * @param {string} command - the action to be performed by the message-adapter
- * @param {Object} input - the input to be sent to the message-adapter
- * @returns {Promise.<Object>} - the output of the message-adapter
+ * @returns {Promise<Object>} cumulusMessageAdapterObject - Returns an Object with a
+ * childprocess and a stderror buffer
+ * @returns {Object} cumulusMessageAdapterObject.cmaProcess - A CMA childProcess Object
+ * @returns {string} cumulusMessageAdapterObject.errorObj -  A Object with the property
+ *                                                           'stderrBuffer' to make the encapsulated
+ *                                                           error event storage outside this method
  */
-async function callCumulusMessageAdapter(command, input) {
-  const spawnArguments = await generateCMASpawnArguments(command);
+async function invokeCumulusMessageAdapter() {
+  const spawnArguments = await generateCMASpawnArguments('stream');
+  const errorObj = { stderrBuffer: '' };
   try {
-    const cumulusMessageAdapter = execa(...spawnArguments);
-    cumulusMessageAdapter.stdin.on('error', () => {});
-    cumulusMessageAdapter.stdin.end(JSON.stringify(input));
-    const { stdout, stderr, exitCode } = await cumulusMessageAdapter;
-    if (exitCode === 0) return (JSON.parse(stdout));
-    throw new CumulusMessageAdapterExecutionError(stderr);
+    // Would like to use sindresorhus's lib, however
+    // https://github.com/sindresorhus/execa/issues/411
+    // and related mean that pulling in the childProcess
+    // is hacky.   Using node native for now, should revisit
+    // the results #414 in the future.
+    const cmaProcess = childProcess.spawn(...spawnArguments);
+    cmaProcess.on('error', () => {});
+    cmaProcess.stdin.setEncoding = 'utf8';
+    cmaProcess.stdout.setEncoding = 'utf8';
+    cmaProcess.stderr.setEncoding = 'utf8';
+    cmaProcess.on('close', () => {
+      console.log(`CMA Exit Code: ${cmaProcess.exitCode} `);
+      if (cmaProcess.exitCode !== 0) {
+        console.log(`CMA Failure: ${errorObj.stderrBuffer}`);
+      }
+    });
+    cmaProcess.stderr.on('data', (data) => {
+      errorObj.stderrBuffer += String(data);
+    });
+    return { cmaProcess, errorObj };
   }
   catch (error) {
     const msg = `CMA process failed (${error.shortMessage})\n
                  Trace: ${error.message}}\n\n\n
-                 STDERR: ${error.stderr}`;
+                 STDERR: ${errorObj.stderrBuffer}`;
     throw new CumulusMessageAdapterExecutionError(msg);
   }
-}
-
-/**
- * If a Cumulus Remote Message is passed, fetch it and return a full Cumulus
- * Message with updated task metadata.
- *
- * @param {Object} cumulusMessage - either a full Cumulus Message or a Cumulus Remote Message
- * @param {Object} context - an AWS Lambda context
- * @param {string} schemaLocations - contains location of schema files, can be null
- * @returns {Promise.<Object>} - a full Cumulus Message
- */
-function loadAndUpdateRemoteEvent(cumulusMessage, context, schemaLocations) {
-  return callCumulusMessageAdapter('loadAndUpdateRemoteEvent', {
-    event: cumulusMessage,
-    context,
-    schemas: schemaLocations
-  });
-}
-
-/**
- * Query AWS to create a task-specific event
- *
- * @param {Object} cumulusMessage - a full Cumulus Message
- * @param {Object} context - an AWS Lambda context
- * @param {string} schemaLocations - contains location of schema files, can be null
- * @returns {Promise.<Object>} - an Object containing the keys input, config and messageConfig
- */
-function loadNestedEvent(cumulusMessage, context, schemaLocations) {
-  return callCumulusMessageAdapter('loadNestedEvent', {
-    event: cumulusMessage,
-    schemas: schemaLocations,
-    context
-  });
-}
-
-/**
- * Create a new Cumulus message with the output of this task
- *
- * @param {Object} handlerResponse - the return value of the task function
- * @param {Object} cumulusMessage - a full Cumulus Message
- * @param {Object} messageConfig - the value of the messageConfig key returned by loadNestedEvent
- * @param {string} schemaLocations - contains location of schema files, can be null
- * @returns {Promise.<Object>} - a Cumulus Message or a Cumulus Remote Message
- */
-function createNextEvent(handlerResponse, cumulusMessage, messageConfig, schemaLocations) {
-  const input = {
-    event: cumulusMessage,
-    handler_response: handlerResponse
-  };
-
-  // If input.message_config is undefined, JSON.stringify will drop the key.
-  // If it is instead set to null, the key is retained and the value is null.
-  input.message_config = messageConfig || null;
-  input.schemas = schemaLocations;
-
-  return callCumulusMessageAdapter('createNextEvent', input);
 }
 
 /**
@@ -232,55 +196,100 @@ function setCumulusEnvironment(cumulusMessage, context) {
 }
 
 /**
+ * Read the CMA output/results after sending a command
+ *
+ * @param {Object} readLine - configured readline object
+ * @param {*} errorObj - cma errorObject with stderr string buffer
+ * @returns {Promise<Object>} - Promise that resolves to a parsed JSON object
+ *                              from the CMA output
+ */
+async function getCmaOutput(readLine, errorObj) {
+  return new Promise((resolve, reject) => {
+    let buffer = '';
+    readLine.resume();
+    readLine.on('line', (input) => {
+      if (input.endsWith('<EOC>')) {
+        readLine.pause();
+        readLine.removeAllListeners('line');
+        const endInput = input.replace('<EOC>', '');
+        buffer += endInput;
+        resolve(JSON.parse(buffer));
+      }
+      else {
+        buffer += input;
+      }
+    });
+    readLine.on('close', () => {
+      reject(new CumulusMessageAdapterExecutionError(errorObj.stderrBuffer));
+    });
+  });
+}
+
+/**
  * Build a nested Cumulus event and pass it to a tasks's business function
  *
  * @param {Function} taskFunction - the function containing the business logic of the task
  * @param {Object} cumulusMessage - either a full Cumulus Message or a Cumulus Remote Message
  * @param {Object} context - an AWS Lambda context
- * @param {Function} callback - the callback to be called when the taskFunction
- *   has completed.  This should be the callback passed to the Lambda handler.
  * @param {string} schemas - Location of schema files, defaults to null.
- * @returns {undefined} - there is no return value from this function, but
- *   the callback function will be invoked with either an error or a full
- *   Cumulus message containing the result of the business logic function.
+ * @returns {Promise<Object>} - The response from the call to createNextEvent or the taskFunction
+ *                     depending on the CUMULUS_MESSAGE_ADAPTER_DISABLED environment variable
  */
-function runCumulusTask(taskFunction, cumulusMessage, context, callback, schemas = null) {
-  let promisedNextEvent;
-  setCumulusEnvironment(cumulusMessage, context);
+async function runCumulusTask(taskFunction, cumulusMessage,
+  context, schemas = null) {
   if (process.env.CUMULUS_MESSAGE_ADAPTER_DISABLED === 'true') {
-    promisedNextEvent = invokePromisedTaskFunction(
+    const functionReturn = await invokePromisedTaskFunction(
       taskFunction,
       cumulusMessage,
       context
     );
+    return functionReturn;
   }
-  else {
-    const promisedRemoteEvent = loadAndUpdateRemoteEvent(cumulusMessage, context, schemas);
-    const promisedNestedEvent = promisedRemoteEvent.then((event) => {
-      // Reset the environment with what we can grab from the event from S3
-      setCumulusEnvironment(event, context);
-      return loadNestedEvent(event, context, schemas);
+  try {
+    const { cmaProcess, errorObj } = await invokeCumulusMessageAdapter();
+    const cmaStdin = cmaProcess.stdin;
+    const rl = readline.createInterface({
+      input: cmaProcess.stdout
     });
+    cmaStdin.write('loadAndUpdateRemoteEvent\n');
+    cmaStdin.write(JSON.stringify({
+      event: cumulusMessage,
+      context,
+      schemas
+    }));
+    cmaStdin.write('\n<EOC>\n');
 
-    const promisedTaskOutput = promisedNestedEvent
-      .then((nestedEvent) => taskFunction(nestedEvent, context));
-
-    promisedNextEvent = Promise.all([promisedTaskOutput, promisedRemoteEvent, promisedNestedEvent])
-      .then((resolvedPromises) => createNextEvent(
-        resolvedPromises[0],
-        resolvedPromises[1],
-        resolvedPromises[2].messageConfig,
-        schemas
-      ));
+    const loadAndUpdateRemoteEventOutput = await getCmaOutput(rl, errorObj);
+    setCumulusEnvironment(loadAndUpdateRemoteEventOutput, context);
+    cmaStdin.write('loadNestedEvent\n');
+    cmaStdin.write(JSON.stringify({
+      event: loadAndUpdateRemoteEventOutput,
+      schemas,
+      context
+    }));
+    cmaStdin.write('\n<EOC>\n');
+    const loadNestedEventOutput = await getCmaOutput(rl, errorObj);
+    const taskOutput = await invokePromisedTaskFunction(taskFunction,
+      loadNestedEventOutput, context);
+    cmaStdin.write('createNextEvent\n');
+    cmaStdin.write(JSON.stringify({
+      event: loadAndUpdateRemoteEventOutput,
+      handler_response: taskOutput,
+      message_config: loadNestedEventOutput.messageConfig,
+      schemas
+    }));
+    cmaStdin.write('\n<EOC>\n');
+    const createNextEventOutput = await getCmaOutput(rl, errorObj);
+    cmaStdin.write('\n<EXIT>\n');
+    return createNextEventOutput;
   }
-
-  promisedNextEvent
-    .then((nextEvent) => callback(null, nextEvent))
-    .catch((err) => {
-      if (err.name && err.name.includes('WorkflowError')) {
-        callback(null, { ...cumulusMessage, payload: null, exception: err.name });
-      }
-      else callback(err);
-    });
+  catch (error) {
+    if (error.name && error.name.includes('WorkflowError')) {
+      return { ...cumulusMessage, payload: null, exception: error.name };
+    }
+    throw error;
+  }
 }
+
 exports.runCumulusTask = runCumulusTask;
+exports.invokeCumulusMessageAdapter = invokeCumulusMessageAdapter;