+++ test/test-index-with-dir-env.js
@@ -20,45 +20,38 @@ test.before(async() => {
   testContext.inputEvent = JSON.parse(fs.readFileSync(inputJson));
   const outputJson = path.join(__dirname, 'fixtures/messages/basic.output.json');
   testContext.outputEvent = JSON.parse(fs.readFileSync(outputJson));
+  process.env.USE_CMA_BINARY = 'true';
 });
 
 test.after.always('final cleanup', () => fs.remove(testContext.dir));
 
-test.cb('CUMULUS_MESSAGE_ADAPTER_DIR sets the location of the message adapter', (t) => {
+test('CUMULUS_MESSAGE_ADAPTER_DIR sets the location of the message adapter', async(t) => {
   const dir = path.join(__dirname, 'alternate-dir', 'cumulus-message-adapter');
   process.env.CUMULUS_MESSAGE_ADAPTER_DIR = dir;
 
   const businessLogicOutput = 42;
-  const businessLogic = () => businessLogicOutput;
+  const businessLogic = async() => businessLogicOutput;
 
   // assign task output from the lambda
   const expectedOutput = clonedeep(testContext.outputEvent);
   expectedOutput.payload = businessLogicOutput;
 
-  function callback(err, data) {
-    t.is(err, null);
-    t.deepEqual(data, expectedOutput);
-    t.end();
-  }
-
-  return cumulusMessageAdapter.runCumulusTask(businessLogic, testContext.inputEvent, {}, callback);
+  const result = await cumulusMessageAdapter.runCumulusTask(
+    businessLogic,
+    testContext.inputEvent,
+    {}
+  );
+  t.deepEqual(result, expectedOutput);
 });
 
-test.cb('callback returns error if CUMULUS_MESSAGE_ADAPTER_DIR is incorrect', (t) => {
+test('callback returns error if CUMULUS_MESSAGE_ADAPTER_DIR is incorrect', async(t) => {
   const dir = path.join(__dirname, 'wrong-dir', 'cumulus-message-adapter');
   process.env.CUMULUS_MESSAGE_ADAPTER_DIR = dir;
 
   const businessLogicOutput = 42;
-  const businessLogic = () => businessLogicOutput;
+  const businessLogic = async() => businessLogicOutput;
 
   const inputEvent = { a: 1 };
 
-  function callback(err, data) {
-    t.is(typeof err, 'object');
-    t.is(err.name, 'CumulusMessageAdapterExecutionError');
-    t.is(data, undefined);
-    t.end();
-  }
-
-  cumulusMessageAdapter.runCumulusTask(businessLogic, inputEvent, {}, callback);
+  await t.throwsAsync(cumulusMessageAdapter.runCumulusTask(businessLogic, inputEvent, {}));
 });